printf() format warrning fix
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@20994 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ samples/dnd/dnd.cpp
@@ -1451,7 +1451,7 @@ bool DnDFile::OnDropFiles(wxCoord, wxCoord, const wxArrayString& filenames)
 {
     size_t nFiles = filenames.GetCount();
     wxString str;
-    str.Printf( _T("%d files dropped"), nFiles);
+    str.Printf( _T("%d files dropped"), (int)nFiles);
     m_pOwner->Append(str);
     for ( size_t n = 0; n < nFiles; n++ ) {
         m_pOwner->Append(filenames[n]);
